moved alloca.hpp to private aux_ and size refactor
